Ver 0.9 Fix some timestamp and index block bug.

+++ YunTable/include/yfile.h
@@ -12,7 +12,7 @@ public YFile* create_new_yfile(char* file_path, ResultSet* resultSet, char* tabl
 
 public ResultSet* query_yfile_by_row_key(YFile* yfile, char* row_key);
 
-public ResultSet* query_yfile_by_timestamp(YFile* yfile, int begin_timestamp, int end_timestamp);
+public ResultSet* query_yfile_by_timestamp(YFile* yfile, long long begin_timestamp, long long end_timestamp);
 
 public void refresh_yfile_data_block_cache(YFile* yfile, int hotness_value);
